feat: overhaul SID playback, expand demo disk, fix WTS pitch tuning
SID playback rewritten to use virtual JSR mechanism — saves/restores
CPU state, zero page, stack, and demo player region so SID routines
can't corrupt the host. Auto-stops after 5 consecutive play hangs.
Relocates SIDs that overlap demo player ($7200+) or hardware ($A000+).
Tolerates invalid BCD digits instead of crashing.

Demo disk updated: removed 5 oversized SIDs (35-56KB), added 17 mono
classics (Hubbard, Galway, Follin, Tel, Dunn, Whittaker, DRAX) and
11 stereo 2SID tunes. Removed RSID/IRQ-only SIDs (arkanoid, myth).
Memory layout reorganized — demo at $7200, keyboard at $9800, giving
27.6KB contiguous SID space.

WavetableSynth pitch fix: SF2 loader now parses coarseTune, fineTune,
and pitchCorrection generators, combined into TuningCents and applied
in the pitch ratio calculation.

+++ e6502UnitTests/NdiDirectoryTests.cs
@@ -24,7 +24,7 @@ public class NdiDirectoryTests
         Assert.AreEqual(NdiFileType.Bas, entry.FileType);
         Assert.AreEqual((ushort)0xFFFF, entry.ParentIndex);
         Assert.AreEqual((ushort)51, entry.StartSector);
-        Assert.AreEqual((ushort)1024, entry.SizeBytes);
+        Assert.AreEqual(1024, entry.SizeBytes);
         Assert.AreEqual((ushort)4, entry.SectorCount);
         Assert.IsTrue(entry.IsActive);
         Assert.IsFalse(entry.IsDirectory);
@@ -121,12 +121,12 @@ public class NdiDirectoryTests
 
         var f1 = rootEntries.First(e => e.Filename == "file1.bas");
         Assert.AreEqual(NdiFileType.Bas, f1.FileType);
-        Assert.AreEqual((ushort)512, f1.SizeBytes);
+        Assert.AreEqual(512, f1.SizeBytes);
         Assert.AreEqual((ushort)2, f1.SectorCount);
 
         var f2 = rootEntries.First(e => e.Filename == "file2.gfx");
         Assert.AreEqual(NdiFileType.Gfx, f2.FileType);
-        Assert.AreEqual((ushort)4096, f2.SizeBytes);
+        Assert.AreEqual(4096, f2.SizeBytes);
 
         Assert.AreEqual("inner.mid", subEntries[0].Filename);
         Assert.AreEqual(NdiFileType.Mid, subEntries[0].FileType);
